+++ lib/dnsimple/registrar.ex
@@ -10,9 +10,11 @@ defmodule Dnsimple.Registrar do
 
   alias Dnsimple.Client
   alias Dnsimple.Response
-  alias Dnsimple.Domain
   alias Dnsimple.DomainCheck
   alias Dnsimple.DomainPremiumPrice
+  alias Dnsimple.DomainRegistration
+  alias Dnsimple.DomainRenewal
+  alias Dnsimple.DomainTransfer
   alias Dnsimple.WhoisPrivacy
   alias Dnsimple.VanityNameServer
 
@@ -83,7 +85,7 @@ defmodule Dnsimple.Registrar do
     url = Client.versioned("/#{account_id}/registrar/domains/#{domain_name}/registration")
 
     Client.post(client, url, attributes, options)
-    |> Response.parse(%{"data" => %Domain{}})
+    |> Response.parse(%{"data" => %DomainRegistration{}})
   end
 
 
@@ -105,7 +107,7 @@ defmodule Dnsimple.Registrar do
     url = Client.versioned("/#{account_id}/registrar/domains/#{domain_name}/renewal")
 
     Client.post(client, url, attributes, options)
-    |> Response.parse(%{"data" => %Domain{}})
+    |> Response.parse(%{"data" => %DomainRenewal{}})
   end
 
 
@@ -131,7 +133,7 @@ defmodule Dnsimple.Registrar do
     url = Client.versioned("/#{account_id}/registrar/domains/#{domain_name}/transfer")
 
     Client.post(client, url, attributes, options)
-    |> Response.parse(%{"data" => %Domain{}})
+    |> Response.parse(%{"data" => %DomainTransfer{}})
   end