+++ test/exvcr_utils_test.exs
@@ -5,23 +5,23 @@ defmodule ExvcrUtilsTest do
     fixture = ExvcrUtils.read_fixture("checkDomain/success.http")
 
     assert fixture == """
-    HTTP/1.1 200 OK
-    Server: nginx
-    Date: Fri, 26 Feb 2016 16:04:05 GMT
-    Content-Type: application/json; charset=utf-8
-    Transfer-Encoding: chunked
-    Connection: keep-alive
-    Status: 200 OK
-    X-RateLimit-Limit: 4000
-    X-RateLimit-Remaining: 3999
-    X-RateLimit-Reset: 1456506245
-    ETag: W/"f3cf8499e935e48401aae26842f712c4"
-    Cache-Control: max-age=0, private, must-revalidate
-    X-Request-Id: e3c6fac1-a36d-42c3-8a04-d68f127add18
-    X-Runtime: 0.605907
-    Strict-Transport-Security: max-age=31536000
-
-    {"data":{"domain":"example.com","available":true,"premium":false}}
+    HTTP/1.1 200 OK\r
+    Server: nginx\r
+    Date: Fri, 26 Feb 2016 16:04:05 GMT\r
+    Content-Type: application/json; charset=utf-8\r
+    Transfer-Encoding: chunked\r
+    Connection: keep-alive\r
+    Status: 200 OK\r
+    X-RateLimit-Limit: 4000\r
+    X-RateLimit-Remaining: 3999\r
+    X-RateLimit-Reset: 1456506245\r
+    ETag: W/"f3cf8499e935e48401aae26842f712c4"\r
+    Cache-Control: max-age=0, private, must-revalidate\r
+    X-Request-Id: e3c6fac1-a36d-42c3-8a04-d68f127add18\r
+    X-Runtime: 0.605907\r
+    Strict-Transport-Security: max-age=31536000\r
+    \r
+    {"data":{"domain":"ruby.codes","available":true,"premium":true}}\
     """
   end